+++ front/src/routes/index.js
@@ -1,12 +1,12 @@
 import { Routes, Route } from 'react-router-dom';
 import Home from '../pages/Home';
-import SignUpForm from '../components/Log/SignUpForm';
+import Authentication from '../pages/Authentication';
 
 function Index() {
     return (
         <Routes>
-            <Route path="/" element={<Home />} />
-            <Route path='/signup' exact="true" element={<SignUpForm />} />
+            <Route path="/" exact="true" element={<Home />} />
+            <Route path="/auth" exact="true" element={<Authentication />} />
         </Routes>
     );
 }
